Combine external example schema corpus tests
